correçoes 1

+++ models/Pdds.js
@@ -14,7 +14,7 @@ const Pdds = db.define('pedidos',{
         allowNull:false
     },
     priceu:{
-        type:DataTypes.FLOAT(5,2),
+        type:DataTypes.FLOAT(6,2),
         allowNull:false
     },
     qtdp:{
@@ -22,7 +22,7 @@ const Pdds = db.define('pedidos',{
         allowNull:false
     },
     vtt_compra:{
-        type:DataTypes.FLOAT(5,2),
+        type:DataTypes.FLOAT(6,2),
         allowNull:false
     }
 })
